requirements.txt
numpy>=0.13
scipy>=1.2.0
scikit-learn>=0.17
tensorflow-gpu==1.12.0
matplotlib
tqdm
tensorboardX
pandas
prefetch_generator
joblib
bintrees
